excel_service/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
dnspython==2.7.0
fastapi==0.115.5
gevent==24.11.1
graphql-core==3.2.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==2.10.0
pydantic_core==2.27.0
pydentic==0.0.1.dev3
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-stdnum==1.20
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
strawberry-graphql==0.252.0
typing_extensions==4.12.2
uvicorn==0.32.1
websockets==14.1
zope.event==5.0
zope.interface==7.1.1
